refactor: move functions for changing locale, getting ContrastingColor, getting color behind navbar to cs file (client_utils); move ss function to server_utils
